Fix for Rom scan failure on some devices
