+++ src/paulscode/android/mupen64plusae/CacheRomInfoFragment.java
@@ -21,7 +21,13 @@
 
 package paulscode.android.mupen64plusae;
 
+import java.io.BufferedOutputStream;
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.zip.ZipEntry;
 
 import org.mupen64plusae.v3.alpha.R;
 
@@ -33,6 +39,7 @@ import paulscode.android.mupen64plusae.task.CacheRomInfoService.CacheRomInfoList
 import paulscode.android.mupen64plusae.task.CacheRomInfoService.LocalBinder;
 import android.app.Activity;
 import android.support.v4.app.Fragment;
+import android.util.Log;
 import android.content.ComponentName;
 import android.content.ServiceConnection;
 import android.os.Bundle;
@@ -203,5 +210,64 @@ public class CacheRomInfoFragment extends Fragment implements CacheRomInfoListen
     public boolean IsInProgress()
     {
         return mInProgress;
+    }
+    
+
+    public static File extractRomFile( File destDir, ZipEntry zipEntry, InputStream inStream )
+    {        
+        // Read the first 4 bytes of the entry
+        byte[] buffer = new byte[1024];
+        try
+        {
+            if( inStream.read( buffer, 0, 4 ) != 4 )
+                return null;
+        }
+        catch( IOException e )
+        {
+            Log.w( "GalleryActivity", e );
+            return null;
+        }
+        
+        // This entry appears to be a valid ROM, extract it
+        Log.i( "GalleryActivity", "Found zip entry " + zipEntry.getName() );
+
+        String entryName = new File( zipEntry.getName() ).getName();
+        File extractedFile = new File( destDir, entryName );
+        try
+        {
+            // Open the output stream (throws exceptions)
+            OutputStream outStream = new FileOutputStream( extractedFile );
+            try
+            {
+                // Buffer the stream
+                outStream = new BufferedOutputStream( outStream );
+                
+                // Write the first four bytes we already peeked at (throws exceptions)
+                outStream.write( buffer, 0, 4 );
+                
+                // Read/write the remainder of the zip entry (throws exceptions)
+                int n;
+                while( ( n = inStream.read( buffer ) ) >= 0 )
+                {
+                    outStream.write( buffer, 0, n );
+                }
+                return extractedFile;
+            }
+            catch( IOException e )
+            {
+                Log.w( "GalleryActivity", e );
+                return null;
+            }
+            finally
+            {
+                // Flush output stream and guarantee no memory leaks
+                outStream.close();
+            }
+        }
+        catch( IOException e )
+        {
+            Log.w( "GalleryActivity", e );
+            return null;
+        }
     }
 }
